route handler and reply route
