+++ views/partials/create-crit.ejs
@@ -1,0 +1,4 @@
+<form action="/crit" method="post">
+    <input type="text" id="create-crit" name="create-crit" >
+    <button id="create-crit-btn">CRIT!</button>
+</form>
