+++ views/partials/menu.ejs
@@ -1,0 +1,17 @@
+<div class="side-menu">
+    <ul>
+        <li>
+            <img src="https://img.icons8.com/material-rounded/30/000000/home-page.png"/>
+            <a href="/static/index.html">Home</a>
+        </li>
+        
+        <li>
+            <img src="https://img.icons8.com/material-rounded/30/000000/user-male-circle.png"/>
+            <a href="/static/profile.html">Profile</a>
+        </li>
+        <li>
+            <img src="https://img.icons8.com/ios-glyphs/30/000000/gear.png"/>
+            <a href="">Settings</a>
+        </li>
+    </ul>
+</div>
