+++ views/partials/search-bar.ejs
@@ -1,0 +1,5 @@
+<div class="search"> 
+    <div>
+        <input id="search" name="search-btn" type="text" placeholder="Search...">
+        <button id="search-btn">Search</button>
+</div>
